refactor: 미니 게임 관련 타입 및 구조 정리
- 카드 게임 관련 타입을 types/miniGame/cardGame.ts로 이동하여 구조를 개선함
- Round 관련 상수를 types/miniGame/round.ts로 이동하고, 기존의 ROUND_NUMBER_MAP을 ROUND_MAP으로 변경, 기존의 CardGameRound을 RoundType으로 변경 (MiniGameTransition에서 사용되는데 카드 도메인이 들어간 타입이 사용되는게 구조상 적절하지 않음)
- MiniGameType 및 관련 타입을 types/miniGame/common.ts로 통합하여 코드의 일관성을 높임
