+++ frontend/src/features/miniGame/components/MiniGameTransition/MiniGameTransition.tsx
@@ -1,12 +1,12 @@
 import Description from '@/components/@common/Description/Description';
 import Headline1 from '@/components/@common/Headline1/Headline1';
 import Layout from '@/layouts/Layout';
-import { CardGameRound, ROUND_NUMBER_MAP } from '@/types/round';
+import { ROUND_MAP, RoundType } from '@/types/miniGame/round';
 import { PropsWithChildren } from 'react';
 import * as S from './MiniGameTransition.styled';
 
 type Props = {
-  currentRound: CardGameRound;
+  currentRound: RoundType;
 } & PropsWithChildren;
 
 const MiniGameTransition = ({ currentRound, children }: Props) => {
@@ -15,7 +15,7 @@ const MiniGameTransition = ({ currentRound, children }: Props) => {
       <S.Container>
         <S.Wrapper>
           <S.DescriptionWrapper>
-            <Headline1 color="white">Round {ROUND_NUMBER_MAP[currentRound]}</Headline1>
+            <Headline1 color="white">Round {ROUND_MAP[currentRound]}</Headline1>
             <Description color="white">다음 라운드로 이동합니다!</Description>
           </S.DescriptionWrapper>
           {children}
