+++ frontend/src/features/miniGame/pages/MiniGamePlayPage/MiniGamePlayPage.tsx
@@ -1,5 +1,5 @@
 import CardGamePlayPage from '@/features/miniGame/cardGame/pages/CardGamePlayPage';
-import { MiniGameType } from '@/types/miniGame';
+import { MiniGameType } from '@/types/miniGame/common';
 import { JSX } from 'react';
 import { useParams } from 'react-router-dom';
