+++ frontend/src/features/miniGame/pages/MiniGameReady/MiniGameReadyPage.tsx
@@ -1,4 +1,4 @@
-import { MiniGameType } from '@/types/miniGame';
+import { MiniGameType } from '@/types/miniGame/common';
 import { JSX } from 'react';
 import { useParams } from 'react-router-dom';
 import CardGameReadyPage from '../../cardGame/pages/CardGameReadyPage';
