+++ frontend/src/features/room/lobby/components/MiniGameSection/MiniGameSection.tsx
@@ -4,7 +4,11 @@ import CardIcon from '@/assets/card-icon.svg';
 import GameActionButton from '@/components/@common/GameActionButton/GameActionButton';
 import SectionTitle from '@/components/@composition/SectionTitle/SectionTitle';
 import { usePlayerType } from '@/contexts/PlayerType/PlayerTypeContext';
-import { MINI_GAME_DESCRIPTION, MINI_GAME_NAME_MAP, MiniGameType } from '@/types/miniGame';
+import {
+  MINI_GAME_DESCRIPTION_MAP,
+  MINI_GAME_NAME_MAP,
+  MiniGameType,
+} from '@/types/miniGame/common';
 import { useEffect, useState } from 'react';
 import * as S from './MiniGameSection.styled';
 
@@ -52,7 +56,7 @@ export const MiniGameSection = ({ selectedMiniGames, handleMiniGameClick }: Prop
             isSelected={selectedMiniGames.includes(miniGame)}
             isDisabled={playerType === 'GUEST'}
             gameName={MINI_GAME_NAME_MAP[miniGame]}
-            description={MINI_GAME_DESCRIPTION[miniGame]}
+            description={MINI_GAME_DESCRIPTION_MAP[miniGame]}
             onClick={() => handleMiniGameClick(miniGame)}
             icon={<S.Icon src={CardIcon} alt={miniGame} />}
           />
