+++ frontend/src/types/miniGame/common.ts
@@ -1,0 +1,14 @@
+/**
+ * 전체 미니 게임 공통 타입
+ */
+
+export const MINI_GAME_NAME_MAP = {
+  CARD_GAME: '카드게임',
+  // '31_GAME': '랜덤 31222',
+} as const;
+
+export type MiniGameType = keyof typeof MINI_GAME_NAME_MAP;
+
+export const MINI_GAME_DESCRIPTION_MAP: Record<MiniGameType, string[]> = {
+  CARD_GAME: ['2라운드 동안 매번 카드 1장씩 뒤집어', '가장 높은 점수를 내보세요!'],
+};
